pyproject.toml
[project]
name = "siftarr"
dynamic = ["version"]
description = "Media search and download decision middleware"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "qbittorrent-api>=2024.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "setuptools-scm>=8.0.0",
]

[build-system]
requires = ["setuptools>=80.0.0", "setuptools-scm>=8.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools_scm]
git_describe_command = "git describe --dirty --tags --long --match v*"
fallback_version = "0.0.0"
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:sugar"

[tool.ruff]
# Ruff configuration - use opinionated defaults
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Enable specific rule sets
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line too long (handled by formatter)
